+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
+- Add `--destination.timeout` and `--diagnostics.timeout` values to set gRPC timeout for
+  primary and diagnostic OTLP Export() calls. (#51)
+
 ## [0.3.0](https://github.com/lightstep/opentelemetry-prometheus-sidecar/releases/tag/v0.3.0) - 2020-12-08
 
 - Change several metric names to use `.` instead of `_` for